Creacion de token y validacion en usuarios y productos
