+++ app.js
@@ -1,14 +1,16 @@
 const express = require("express");
 const morgan = require("morgan");
-const productsController = require("./controllers/productsController.js");
-const usersController = require("./controllers/usersController.js");
-const { PORT } = require("./db/config.js");
+const productsController = require("./src/controllers/productsController.js");
+const usersController = require("./src/controllers/usersController.js");
+const loginController = require ("./src/controllers/loginController.js")
+const { PORT } = require("./src/db/config.js");
+const validateToken = require("./src/middlewares/validateToken.js");
 
 const app = express();
 
 // Middleware para habilitar CORS
 app.use((req, res, next) => {
-  res.header("Access-Control-Allow-Origin", "*"); // Reemplaza * con el dominio de tu aplicación si es necesario
+  res.header("Access-Control-Allow-Origin", "*");
   res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
   next();
 });
@@ -22,8 +24,10 @@ app.use(morgan("dev"));
 app.use(express.json());
 
 // Se manda a llamar al controlador de catálogo productos
-app.use('/products', productsController);
+app.use('/products',validateToken, productsController);
 
 // Se manda a llamar al controlador del usuario
-app.use('/users', usersController);
+app.use('/users',validateToken, usersController);
 
+//Se manda a llamar al login
+app.use('/login', loginController);
