--- db/config.js
@@ -1,15 +1,0 @@
-const PORT = process.env.PORT || 4000;
-const DB_HOST = process.env.DB_HOST || 'localhost';
-const DB_NAME = process.env.DB_NAME || 'e-commerce';
-const DB_USER = process.env.DB_USER || 'root';
-const DB_PASSWORD = process.env.DB_PASSWORD || '';
-const DB_PORT = process.env.DB_PORT || 3306;
-
-module.exports = {
-    PORT,
-    DB_HOST,
-    DB_NAME,
-    DB_USER,
-    DB_PASSWORD,
-    DB_PORT
-};
